proper unzipping - damn zip standard not supporting non-ascii, damn traum

+++ core/src/ru/alepar/ebook/convert/CalibreConverter.java
@@ -1,10 +1,14 @@
 package ru.alepar.ebook.convert;
 
+import org.apache.commons.compress.archivers.zip.ZipArchiveEntry;
+import org.apache.commons.compress.archivers.zip.ZipFile;
 import ru.alepar.ebook.format.EbookType;
 import ru.alepar.ebook.format.FormatProvider;
+import ru.alepar.ebook.format.StaticFormatProvider;
+import ru.alepar.io.IOUtils;
 import ru.alepar.lib.stuff.Oops;
 
-import java.io.File;
+import java.io.*;
 
 public class CalibreConverter implements Converter {
 
@@ -20,11 +24,17 @@ public class CalibreConverter implements Converter {
 
     @Override
     public File convertFor(EbookType type, File file) {
+        if (type == EbookType.UNKNOWN) {
+            return file; // ebook not recognized, conversion will not be performed
+        }
         try {
+            if (file.getName().endsWith(".zip")) {
+                file = uncompress(file);
+            }
             String inputPath = file.getCanonicalPath();
             File outputFile = File.createTempFile("libby", '.' + provider.extension(type));
             String outputFileName = outputFile.getCanonicalPath();
-            if (!outputFile.delete()) {
+            if (!outputFile.delete()) { // don't need the file itself, only it's name
                 throw new Oops("cudnt trash " + outputFileName);
             }
             int retCode = exec.exec(new String[]{
@@ -41,6 +51,38 @@ public class CalibreConverter implements Converter {
         } catch (Exception e) {
             throw new RuntimeException("conversion failed", e);
         }
+    }
+
+    private File uncompress(File inFile) throws IOException {
+        String fileName = inFile.getName();
+        String fileNameWithoutZipExtension = fileName.substring(0, fileName.length() - 4);
+        ZipFile zipFile = new ZipFile(inFile);
+        ZipArchiveEntry entry = getFirstEntry(zipFile);
+        InputStream is = zipFile.getInputStream(entry);
+        try {
+            File outFile = File.createTempFile("libby", fileNameWithoutZipExtension);
+            OutputStream fos = new FileOutputStream(outFile);
+            try {
+                IOUtils.copy(is, fos);
+                return outFile;
+            } finally {
+                fos.close();
+            }
+        } finally {
+            is.close();
+        }
+    }
+
+    private ZipArchiveEntry getFirstEntry(ZipFile zipFile) {
+        return (ZipArchiveEntry) zipFile.getEntriesInPhysicalOrder().nextElement();
+    }
+
+    public static void main(String[] args) throws Exception {
+        String binary = "c:\\Program Files (x86)\\Calibre2\\ebook-convert.exe";
+        CalibreConverter converter = new CalibreConverter(binary, new RuntimeExec(), new StaticFormatProvider());
+
+        File out = converter.convertFor(EbookType.KINDLE_DX, new File("c:\\temp\\Лукьяненко 1 Геном.fb2.zip"));
+        System.out.println(out.getCanonicalPath());
     }
 
 }
